chore: Add Jest UUID matcher extension
This commit adds a new file `jest_uuidmatcher.js` to the `extensions` directory. The file contains a custom Jest matcher `toBeUUID` that checks if a received value is a valid UUID. The matcher uses a regular expression to validate the UUID format.

This extension enhances the testing capabilities of the todo app by providing a convenient way to assert UUID values in tests.