feat: Add task completion functionality
This commit adds the ability to mark a task as done in the TodoManager class. A new method `markAsDone` is implemented, which takes the task ID as a parameter and updates the `done` property of the corresponding task to `true`. This functionality allows users to track the completion status of tasks in the todo app.