feat: Add task search functionality
This commit introduces the ability to search for tasks by title in the TodoManager class. A new method `searchTasksByTitle` is implemented, which takes a title as a parameter and returns an array of tasks that match the search criteria. This feature enhances the usability of the todo app by allowing users to easily find specific tasks based on their titles.